Elements
This is a fairly substantial update; the app now allows flexible
elemental inputs. Before, it eliminated any sources if they did not
have corresponding elemental data to the user inputted data. Now, the
sourcing function automatically eliminates these elements from the
fingerprinting algorithm, if the user chooses to do this.

This does make the results more vulnerable to bad source matching
though, so it is defaulted off unless the user opts into it.

The database has also been updated.